chore:refactor and code clean-up
